requirements.txt
asgiref==3.3.1
Django==3.1.7
django-js-asset==1.2.2
django-location-field==2.1.0
django-mptt==0.11.0
djangorestframework==3.12.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
